bot/requirements.txt
python-telegram-bot==21.6
feedparser==6.0.11
python-dotenv==1.0.1
aiohttp==3.10.10
tweepy>=4.14.0
pyperclip>=1.8.0
openai>=1.30.0
atproto>=0.0.50
Mastodon.py>=1.8.0
yfinance>=0.2.40
